Minor fixes in the messaging protocol

+++ common/src/main/scala/messageProtocols/SolarPanelData.scala
@@ -5,7 +5,7 @@ class SolarPanelData(val timestamp: Long, // will be part of the key for joining
                      val location: String, // will be part of the key for joining
                      val panelId: String, // will NOT be part of the key for joining weather stream
                      val plantId: String, // will NOT be part of the key for joining
-                     val panelVoltage: Float, // for performance monitoring
-                     val panelCurrent: Float,  // for performance monitoring
-                     val panelTemperature: Float // for predictive maintenance / malfunction mitigation
+                     val panelVoltage: Double, // for performance monitoring
+                     val panelCurrent: Double,  // for performance monitoring
+                     val panelTemperature: Double // for predictive maintenance / malfunction mitigation
                     ) {}
